pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "project page_analyzer"
authors = ["Dmitriy G. <llogginn@gmail.com>"]
repository = "https://github.com/braimm/python-project-83"
readme = "README.md"
packages = [{ include = "page_analyzer" },]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.28.3"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.15.0"
types-psycopg2 = "^2.9.21.20250121"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
